refactor(Scheduling): extract PoolsSuggest to a separate file [YTFRONT-4075]
