+++ packages/ui/src/ui/pages/scheduling/PoolsSuggest/PoolsSuggest.scss
@@ -1,0 +1,5 @@
+.yt-pools-suggest {
+    &__popup {
+        min-width: 220px;
+    }
+}
