+++ packages/ui/src/ui/pages/scheduling/Scheduling/SchedulingTopRowContent.tsx
@@ -8,11 +8,11 @@ import {Breadcrumbs, BreadcrumbsItem} from '@gravity-ui/uikit';
 import ErrorBoundary from '../../../components/ErrorBoundary/ErrorBoundary';
 import {RowWithName} from '../../../containers/AppNavigation/TopRowContent/SectionName';
 import Favourites from '../../../components/Favourites/Favourites';
+import {PoolsSuggest} from '../PoolsSuggest/PoolsSuggest';
 import {
     getCurrentPool,
     getCurrentPoolPath,
     getPool,
-    getPoolsNames,
     getTree,
     getTreesSelectItems,
 } from '../../../store/selectors/scheduling/scheduling';
@@ -22,7 +22,6 @@ import {
     changeTree,
     togglePoolFavourites,
 } from '../../../store/actions/scheduling/scheduling';
-import {schedulingLoadFilterAttributes} from '../../../store/actions/scheduling/scheduling-ts';
 
 import {ROOT_POOL_NAME, SCHEDULING_ALLOWED_ROOT_TABS, Tab} from '../../../constants/scheduling';
 import ClipboardButton from '../../../components/ClipboardButton/ClipboardButton';
@@ -31,7 +30,6 @@ import {Page} from '../../../constants';
 import Link from '../../../components/Link/Link';
 import {EditButton} from '../../../components/EditableAsText/EditableAsText';
 import Select from '../../../components/Select/Select';
-import Suggest from '../../../components/Suggest/Suggest';
 import CreatePoolButton from '../Instruments/CreatePoolDialog/CreatePoolDialog';
 
 import {makeRoutedURL} from '../../../store/location';
@@ -211,69 +209,6 @@ function EditableSchedulingBreadcrumbs() {
                 </React.Fragment>
             )}
         </div>
-    );
-}
-
-function PoolsSuggest(props: {onCancelEdit: () => void}) {
-    const {onCancelEdit} = props;
-    const poolNames = useSelector(getPoolsNames);
-    const tree = useSelector(getTree);
-    const dispatch = useDispatch();
-
-    const getSuggestItems = React.useCallback(
-        (_items: any, filter?: string) => {
-            if (!filter) {
-                return poolNames;
-            }
-
-            const match: Array<string> = [];
-            const startsWith: Array<string> = [];
-            const filtered: Array<string> = [];
-
-            const lcFilter = filter?.toLowerCase();
-
-            _.forEach(poolNames, (poolName) => {
-                const lcPoolName = poolName.toLowerCase();
-                if (lcFilter === lcPoolName) {
-                    match.push(poolName);
-                } else if (lcPoolName.startsWith(lcFilter)) {
-                    startsWith.push(poolName);
-                } else if (poolName !== ROOT_POOL_NAME && -1 !== lcPoolName.indexOf(lcFilter)) {
-                    filtered.push(poolName);
-                }
-            });
-            return match.concat(startsWith, filtered);
-        },
-        [poolNames],
-    );
-
-    const handleCancelEdit = React.useCallback(() => {
-        setTimeout(onCancelEdit, 500);
-    }, [onCancelEdit]);
-
-    const onItemClick = React.useCallback(
-        (pool: string) => {
-            dispatch(changePool(pool));
-            onCancelEdit();
-        },
-        [dispatch, onCancelEdit],
-    );
-
-    const onFocus = React.useCallback(() => {
-        dispatch(schedulingLoadFilterAttributes(tree));
-    }, [dispatch, tree]);
-
-    return (
-        <Suggest
-            popupClassName={block('pool-suggest-popup')}
-            autoFocus
-            filter={getSuggestItems}
-            onBlur={handleCancelEdit}
-            onFocus={onFocus}
-            placeholder="Select pool..."
-            onItemClick={(item) => onItemClick('string' === typeof item ? item : item.value)}
-            items={poolNames}
-        />
     );
 }
